LPS-58241 move upgrade for asset publisher to the module
